+++ include/modules/ipc.hpp
@@ -54,12 +54,17 @@ namespace modules {
     bool has_hook() const;
 
     void set_hook(int h);
-
+    void update_output() ;
    private:
-    static constexpr const char* TAG_OUTPUT{"<output>"};
+    static constexpr auto TAG_OUTPUT = "<output>";
+    static constexpr auto TAG_LABEL = "<label>";
+
+    label_t m_label;
+
     vector<unique_ptr<hook>> m_hooks;
     map<mousebtn, string> m_actions;
     string m_output;
+
     int m_initial{-1};
     int m_current_hook{-1};
     void exec_hook();